Generic version of the validateIndexedStates method.
