+++ rspace/src/test/scala/coop/rchain/rspace/HistoryActionsTests.scala
@@ -47,48 +47,48 @@ trait HistoryActionsTests
       }
     }
 
-  case class State(
-      checkpoint: Blake2b256Hash,
-      contents: Map[Seq[String], Row[Pattern, String, StringsCaptor]],
-      joins: Map[Blake2b256Hash, Seq[Seq[String]]]
-  )
+//  case class State(
+//      checkpoint: Blake2b256Hash,
+//      contents: Map[Seq[String], Row[Pattern, String, StringsCaptor]],
+//      joins: Map[Blake2b256Hash, Seq[Seq[String]]]
+//  )
 
-  def validateIndexedStates(space: ISpace[String, Pattern, String, String, StringsCaptor],
-                            indexedStates: Seq[(State, Int)]): Boolean = {
-    val tests: Seq[Any] = indexedStates
-      .map {
-        case (State(checkpoint, expectedContents, expectedJoins), chunkNo) =>
-          space.reset(checkpoint)
-          val num = "%02d".format(chunkNo)
-
-          val contentsTest = space.store.toMap == expectedContents
-
-          if (contentsTest) {
-            logger.debug(s"$num: store had expected contents")
-          } else {
-            logger.error(s"$num: store had unexpected contents")
-          }
-
-          val actualJoins = space.store.joinMap
-
-          val joinsTest =
-            expectedJoins.forall {
-              case (hash: Blake2b256Hash, expecteds: Seq[Seq[String]]) =>
-                val expected = HashMultiset.create[Seq[String]](expecteds.asJava)
-                val actual   = HashMultiset.create[Seq[String]](actualJoins(hash).asJava)
-                expected.equals(actual)
-            }
-
-          if (joinsTest) {
-            logger.debug(s"$num: store had expected joins")
-          } else {
-            logger.error(s"$num: store had unexpected joins")
-          }
-
-          contentsTest && joinsTest
-      }
-    !tests.contains(false)
-  }
+//  def validateIndexedStates(space: ISpace[String, Pattern, String, String, StringsCaptor],
+//                            indexedStates: Seq[(State, Int)]): Boolean = {
+//    val tests: Seq[Any] = indexedStates
+//      .map {
+//        case (State(checkpoint, expectedContents, expectedJoins), chunkNo) =>
+//          space.reset(checkpoint)
+//          val num = "%02d".format(chunkNo)
+//
+//          val contentsTest = space.store.toMap == expectedContents
+//
+//          if (contentsTest) {
+//            logger.debug(s"$num: store had expected contents")
+//          } else {
+//            logger.error(s"$num: store had unexpected contents")
+//          }
+//
+//          val actualJoins = space.store.joinMap
+//
+//          val joinsTest =
+//            expectedJoins.forall {
+//              case (hash: Blake2b256Hash, expecteds: Seq[Seq[String]]) =>
+//                val expected = HashMultiset.create[Seq[String]](expecteds.asJava)
+//                val actual   = HashMultiset.create[Seq[String]](actualJoins(hash).asJava)
+//                expected.equals(actual)
+//            }
+//
+//          if (joinsTest) {
+//            logger.debug(s"$num: store had expected joins")
+//          } else {
+//            logger.error(s"$num: store had unexpected joins")
+//          }
+//
+//          contentsTest && joinsTest
+//      }
+//    !tests.contains(false)
+//  }
 
   "createCheckpoint on an empty store" should "return the expected hash" in withTestSpace { space =>
     space.createCheckpoint().root shouldBe Blake2b256Hash.fromHex(
@@ -379,7 +379,7 @@ trait HistoryActionsTests
             (State(space.createCheckpoint().root, space.store.toMap, space.store.joinMap), chunkNo)
         }
 
-        validateIndexedStates(space, states)
+        validateIndexedStates(space, states, "3", false)
       }
     }
     check(prop)
@@ -403,7 +403,7 @@ trait HistoryActionsTests
             (State(space.createCheckpoint().root, space.store.toMap, space.store.joinMap), chunkNo)
         }
 
-        validateIndexedStates(space, states)
+        validateIndexedStates(space, states, "1", false)
       }
     }
     check(prop)
@@ -432,7 +432,7 @@ trait HistoryActionsTests
             (State(space.createCheckpoint().root, space.store.toMap, space.store.joinMap), chunkNo)
         }
 
-        validateIndexedStates(space, states)
+        validateIndexedStates(space, states, "2", false)
       }
     }
     check(prop)
